backend/requirements.txt
fastapi
strawberry-graphql
uvicorn[standard]
